+++ test/td_dq_web/controllers/rule_controller_test.exs
@@ -5,19 +5,50 @@ defmodule TdDqWeb.RuleControllerTest do
   alias TdCache.{Audit, Redix}
   alias TdDq.Rules.Rule
 
-  setup do
+  setup_all do
+    domain = CacheHelpers.insert_domain()
+    [domain: domain]
+  end
+
+  setup tags do
     start_supervised!(TdDq.MockRelationCache)
     start_supervised!(TdDd.Search.MockIndexWorker)
     start_supervised!(TdDq.Cache.RuleLoader)
     on_exit(fn -> Redix.del!(Audit.stream()) end)
-
-    [rule: insert(:rule)]
+    domain_id = get_in(tags, [:domain, :id])
+    [rule: insert(:rule, domain_id: domain_id)]
   end
 
   describe "index" do
     @tag authentication: [role: "admin"]
     test "lists all rules", %{conn: conn, swagger_schema: schema} do
       assert %{"data" => [_rule]} =
+               conn
+               |> get(Routes.rule_path(conn, :index))
+               |> validate_resp_schema(schema, "RulesResponse")
+               |> json_response(:ok)
+    end
+
+    @tag authentication: [role: "admin"]
+    test "lists all rules with preloaded domains", %{
+      conn: conn,
+      domain: %{id: domain_id, external_id: external_id, name: name},
+      rule: %{id: rule_id},
+      swagger_schema: schema
+    } do
+      assert %{
+               "data" => [
+                 %{
+                   "id" => ^rule_id,
+                   "domain_id" => ^domain_id,
+                   "domain" => %{
+                     "id" => ^domain_id,
+                     "name" => ^name,
+                     "external_id" => ^external_id
+                   }
+                 }
+               ]
+             } =
                conn
                |> get(Routes.rule_path(conn, :index))
                |> validate_resp_schema(schema, "RulesResponse")
@@ -39,10 +70,11 @@ defmodule TdDqWeb.RuleControllerTest do
       claims: %{user_id: user_id},
       swagger_schema: schema
     } do
-      %{id: id} = insert(:rule, business_concept_id: "42")
+      business_concept_id = Integer.to_string(System.unique_integer([:positive]))
+      %{id: id} = insert(:rule, business_concept_id: business_concept_id)
       insert(:rule, business_concept_id: "1234")
 
-      create_acl_entry(user_id, "business_concept", "42", [:view_quality_rule])
+      create_acl_entry(user_id, "business_concept", business_concept_id, [:view_quality_rule])
 
       assert %{"data" => data} =
                conn
@@ -54,12 +86,129 @@ defmodule TdDqWeb.RuleControllerTest do
     end
   end
 
+  describe "get rule" do
+    @tag authentication: [role: "admin"]
+    test "gets rule by id", %{conn: conn, swagger_schema: schema} do
+      %{id: id, name: name} = insert(:rule)
+
+      assert %{"data" => %{"id" => ^id, "name" => ^name}} =
+               conn
+               |> get(Routes.rule_path(conn, :show, id))
+               |> validate_resp_schema(schema, "RuleResponse")
+               |> json_response(:ok)
+    end
+
+    @tag authentication: [role: "admin"]
+    test "gets rule by id with enriched domain", %{
+      conn: conn,
+      domain: %{id: domain_id, external_id: external_id, name: domain_name},
+      swagger_schema: schema
+    } do
+      %{id: id, name: name} = insert(:rule, domain_id: domain_id)
+
+      assert %{
+               "data" => %{
+                 "id" => ^id,
+                 "name" => ^name,
+                 "domain_id" => ^domain_id,
+                 "domain" => %{
+                   "id" => ^domain_id,
+                   "name" => ^domain_name,
+                   "external_id" => ^external_id
+                 }
+               }
+             } =
+               conn
+               |> get(Routes.rule_path(conn, :show, id))
+               |> validate_resp_schema(schema, "RuleResponse")
+               |> json_response(:ok)
+    end
+
+    @tag authentication: [role: "user"]
+    test "unauthorized when user has no permissions", %{
+      conn: conn
+    } do
+      %{id: id} = insert(:rule)
+
+      assert %{"errors" => %{"detail" => "Forbidden"}} =
+               conn
+               |> get(Routes.rule_path(conn, :show, id))
+               |> json_response(:forbidden)
+    end
+
+    @tag authentication: [user_name: "not_an_admin"]
+    test "gets rule when user has permissions", %{
+      conn: conn,
+      claims: %{user_id: user_id},
+      swagger_schema: schema
+    } do
+      business_concept_id = Integer.to_string(System.unique_integer([:positive]))
+      %{id: id, name: name} = insert(:rule, business_concept_id: business_concept_id)
+      create_acl_entry(user_id, "business_concept", business_concept_id, [:view_quality_rule])
+
+      %{"data" => %{"id" => ^id, "name" => ^name}} =
+        conn
+        |> get(Routes.rule_path(conn, :show, id))
+        |> validate_resp_schema(schema, "RuleResponse")
+        |> json_response(:ok)
+    end
+  end
+
   describe "get_rules_by_concept" do
     @tag authentication: [role: "admin"]
     test "lists all rules of a concept", %{conn: conn, swagger_schema: schema} do
-      assert %{"data" => []} =
+      business_concept_id = Integer.to_string(System.unique_integer([:positive]))
+
+      %{id: id1, business_concept_id: business_concept_id} =
+        insert(:rule, business_concept_id: business_concept_id)
+
+      %{id: id2} = insert(:rule, business_concept_id: business_concept_id)
+
+      assert %{
+               "data" => [
+                 %{"id" => ^id1, "business_concept_id" => ^business_concept_id},
+                 %{"id" => ^id2, "business_concept_id" => ^business_concept_id}
+               ]
+             } =
                conn
-               |> get(Routes.rule_path(conn, :get_rules_by_concept, "id"))
+               |> get(Routes.rule_path(conn, :get_rules_by_concept, business_concept_id))
+               |> validate_resp_schema(schema, "RulesResponse")
+               |> json_response(:ok)
+    end
+
+    @tag authentication: [role: "admin"]
+    test "lists all rules of a concept with entiched domains", %{
+      conn: conn,
+      domain: %{id: domain_id, external_id: external_id, name: name},
+      swagger_schema: schema
+    } do
+      business_concept_id = Integer.to_string(System.unique_integer([:positive]))
+      %{id: id1} = insert(:rule, domain_id: domain_id, business_concept_id: business_concept_id)
+
+      %{id: id2} = insert(:rule, domain_id: domain_id, business_concept_id: business_concept_id)
+
+      assert %{
+               "data" => [
+                 %{
+                   "id" => ^id1,
+                   "domain" => %{
+                     "id" => ^domain_id,
+                     "name" => ^name,
+                     "external_id" => ^external_id
+                   }
+                 },
+                 %{
+                   "id" => ^id2,
+                   "domain" => %{
+                     "id" => ^domain_id,
+                     "name" => ^name,
+                     "external_id" => ^external_id
+                   }
+                 }
+               ]
+             } =
+               conn
+               |> get(Routes.rule_path(conn, :get_rules_by_concept, business_concept_id))
                |> validate_resp_schema(schema, "RulesResponse")
                |> json_response(:ok)
     end
@@ -78,8 +227,8 @@ defmodule TdDqWeb.RuleControllerTest do
 
   describe "create rule" do
     @tag authentication: [role: "admin"]
-    test "renders rule when data is valid", %{conn: conn, swagger_schema: schema} do
-      params = string_params_for(:rule)
+    test "renders rule when data is valid", %{conn: conn, swagger_schema: schema, domain: domain} do
+      params = string_params_for(:rule, domain_id: domain.id)
 
       assert %{"data" => data} =
                conn
@@ -93,10 +242,11 @@ defmodule TdDqWeb.RuleControllerTest do
     @tag authentication: [role: "admin"]
     test "renders rule when data is valid without business concept", %{
       conn: conn,
+      domain: domain,
       swagger_schema: schema
     } do
       rule_params =
-        string_params_for(:rule)
+        string_params_for(:rule, domain_id: domain.id)
         |> Map.delete("business_concept_id")
 
       assert %{"data" => data} =
@@ -148,8 +298,9 @@ defmodule TdDqWeb.RuleControllerTest do
   end
 
   describe "update rule" do
-    setup do
-      [rule: insert(:rule)]
+    setup tags do
+      domain_id = get_in(tags, [:domain, :id])
+      [rule: insert(:rule, domain_id: domain_id)]
     end
 
     @tag authentication: [role: "admin"]